requirements.txt
streamlit
numpy
Pillow
extcolors
scipy
rembg
opencv-python